requirements.txt
langchain
openai
sentence-transformers
faiss-cpu
pymupdf
streamlit
python-dotenv
